Feat: Dynamic styles
This commit adds the feature that allows one to include
dynamic effects, such as hover and pressure, to the
style options. The user is able to add his or her own
effects using the JS events. The syntax is closely linked
to the original Styles library.

I have added the comments, but they do not seem to compile
correctly. I would really appreciate some help with that.
Thank you.